refactor(commands): share accelerator helpers and awaitTrue
Extract the keybinding-token normalization that desktopCommandSurface.ts
owned into src/shared/commands/accelerators.ts so any host (palette UI,
future CLI, web shell) can format catalog keybindings without redefining
the cmd/ctrl/alt/shift translation rules. The desktop surface now imports
toElectronAccelerator, toPlatformAccelerator, and formatDesktopAccelerator
from shared and re-exports the public ones for existing call sites.

Also move the awaitTrue helper from extensionCommandSurface.ts (and its
duplicate in ExtensionCommandRegistry.vitest.ts) into the shared registry
module beside dispatchCommandFromRegistry, where the rejection-propagation
guarantee it enforces is documented. Both consumers now import the single
copy.
